requirements.txt
pipecat-ai[cartesia,openai,silero,deepgram]
fastapi
uvicorn
python-dotenv
loguru
google-genai>=1.0.0
google-cloud-speech>=2.26.0
google-cloud-texttospeech>=2.16.0
twilio
boto3
